# perf(io): excise last use of `std::regex`