On my system, with a GCC 13 Debug build of Cppfront,
lowering `passthrough-tests/clang-12-libstdc++-e.cpp2`
is reduced from 24 s to 0.4 s.
Lowering the other passthrough-tests also takes 0.4 s each.

The passthrough-tests pass, as do the regression-tests:
```output
100% tests passed, 0 tests failed out of 660

Total Test time (real) =  85.42 sec
```