@hsutter 2023-06-17
Holy carruthers, Batman, that's impressive. Between this and the other regex excision you did yesterday, that's a 95+% speedup in cppfront all-up... simply by not using `<regex>`.

I am quickly gaining sympathy with the people who've been complaining about regex perf for many years now and recommend against using it. I've never felt the pain because I haven't been a big regex user, and was I trying regex here in cppfront because I felt I should. Now I won't. Thanks.

One small fix though: You need to change 

    *line.end()

to

    line.data()+line.length()

in two places, because the first is UB (and the Microsoft STL detects it and fires an assertion on it).