@hsutter 2023-06-17
For perspective, I knew compiling the `/passthrough-tests` was kind of slow, because they're huge files that contain entire preprocessed commercial STL implementations. I hadn't bothered to try to find the reason why because I thought I’d put that off until seeing whether the project gets to a point where there’s actual demand to process huge files. And even then, I suspected the culprit was probably iostreams.

Nope, iostreams perf is just fine. It was regex all along.

Just goes to show, as always: Never guess, measure.

Thanks again.