@JohelEGP 2023-06-17
> because the first is UB

That makes sense.
`data() + size()` is guaranteed to be `char{}`,
so I thought it'd be OK to dereference the past-the-end iterator.
Thank you.