@hsutter 2023-06-17
The key is that `data() + size()` avoids dereferencing the one-past-the-end pointer. Even though `&*end()` doesn't use the dereferenced value, it (ever so briefly) derefs the one-past-the-end pointer which triggers the UB that MS-STL diagnoses.