@JohelEGP 2023-06-18
It occurred to me that this could be refactored to behave like [`do_is_keyword`](https://github.com/hsutter/cppfront/blob/e0a6c34d9c102e6045319e92029238cdeb9c72a5/source/lex.h#L936-L952).
To do a single lazy pass and then check that a space follows the `import` keyword.
Unfortunately, I was occupied and you were faster to merge.