@hsutter 2023-06-18
No worries, feel free to do a new PR to do it that way instead.